64-bit-related fixes

+++ src/io/MemReadWriter.cc
@@ -24,7 +24,8 @@ std::pair<Sirikata::uint32, Sirikata::JpegError> MemReadWriter::Read(Sirikata::u
     //fprintf(stderr, "%d READ %02x%02x%02x%02x - %02x%02x%02x%02x\n", (uint32)actualBytesRead, data[0], data[1],data[2], data[3],
     //        data[actualBytesRead-4],data[actualBytesRead-3],data[actualBytesRead-2],data[actualBytesRead-1]);
 
-    std::pair<Sirikata::uint32, JpegError> retval(actualBytesRead, err);
+	//	size_t -> unsigned int is safe because size is unsigned int
+	std::pair<Sirikata::uint32, JpegError> retval(static_cast<unsigned int>(actualBytesRead), err);
     return retval;
 }
 }
